BoutiqueDAL/Infrastructure/Implementations/Database/Boutique/Table/Carts/CartItemTable.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Linq.Expressions;
using BoutiqueCommon.Models.Domain.Interfaces.Carts;
using BoutiqueCommon.Models.Domain.Interfaces.Clothes.ClothesDomains;
using BoutiqueDAL.Infrastructure.Implementations.Database.Base.EntityDatabaseTable;
using BoutiqueDAL.Infrastructure.Interfaces.Database.Boutique.Table.Carts;
using BoutiqueDAL.Infrastructure.Interfaces.Database.Boutique.Table.Clothes;
using BoutiqueDAL.Models.Implementations.Entities.Carts;
using BoutiqueDAL.Models.Implementations.Entities.Clothes;
using BoutiqueDAL.Models.Interfaces.Entities.Carts;
using Microsoft.EntityFrameworkCore;

namespace BoutiqueDAL.Infrastructure.Implementations.Database.Boutique.Table.Carts
{
    /// <summary>
    /// Таблица базы данных позиций корзин
    /// </summary>
    public class CartItemTable : EntityDatabaseTable<Guid, ICartItemDomain, CartItemEntity>, ICartItemTable
    {
        public CartItemTable(DbSet<CartItemEntity> cartItemSet)
            : base(cartItemSet)
        { }

        /// <summary>
        /// Выгрузка идентификатора
        /// </summary>
        public override Expression<Func<CartItemEntity, Guid>> IdSelect() =>
            entity => entity.Id;

        /// <summary>
        /// Функция поиска по идентификатору
        /// </summary>
        public override Expression<Func<CartItemEntity, bool>> IdPredicate(Guid id) =>
            entity => entity.Id == id;

        /// <summary>
        /// Функция поиска по параметрам
        /// </summary>
        public override Expression<Func<CartItemEntity, bool>> IdsPredicate(IEnumerable<Guid> ids) =>
            entity => ids.Contains(entity.Id);
    }
}